Consolidate Prodigy annotation recipes in a single file
